requirements.txt
# torch>=2.0.0
# torchvision>=0.15.0
# matplotlib>=3.5.0
# numpy>=1.21.0
# scikit-learn>=1.0.2
# seaborn>=0.11.2
# pillow>=9.0.0
# opencv-python>=4.6.0.66
# streamlit>=1.10.0
# gdown>=4.4.0


# Core Libraries
# numpy
# pandas
# matplotlib
# opencv-python
# Pillow

# # Deep Learning
# torch>=1.12
# torchvision>=0.13

# # Streamlit App
# streamlit>=1.22

# # Utility
# tqdm
# scikit-learn

# # For Grad-CAM visualization
# seaborn

# # Optional: If using HuggingFace datasets or gdown
# requests

# # Core libraries
# numpy>=1.24.0
# pandas>=1.5.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
# opencv-python>=4.7.0
# Pillow>=9.5.0

# # TensorFlow and related tools
# tensorflow>=2.12.0
# tensorflow-addons>=0.20.0
# tf-explain>=0.3.1

# # Model evaluation
# scikit-learn>=1.2.0

# # Streamlit app
# streamlit>=1.23.0

# # Utilities
# tqdm>=4.65.0

# tensorflow>=2.13
# keras-cv>=0.6.0
# numpy>=1.23
# matplotlib>=3.7
# scikit-learn>=1.3
# opencv-python>=4.8
# streamlit>=1.28
# Pillow>=9.5
# keras-cv-attention-models>=0.20.0

# requirements.txt

tensorflow>=2.15.0
scikit-learn>=1.4.2
numpy>=1.26.4
opencv-python>=4.9.0.80
Pillow>=10.3.0
matplotlib>=3.8.4
streamlit>=1.35.0
keras>=2.15.0
tensorflow-addons>=0.22.0
tensorflow-hub>=0.16.1
h5py>=3.10.0
